refactor: split settings.rs into settings/mod.rs + settings/input.rs
