+++ src/window/settings/mod.rs
@@ -1,8 +1,6 @@
-use crate::core::config::{APP_AUTHOR, APP_HOMEPAGE, APP_VERSION, AppConfig, DockPosition};
-use crate::core::i18n::{current_lang, init_i18n, set_lang, tr};
-use crate::core::persistence::save_config;
+use crate::core::config::{APP_AUTHOR, APP_VERSION, AppConfig, DockPosition};
+use crate::core::i18n::{current_lang, tr};
 use crate::utils::anim::AnimPool;
-use crate::utils::autostart::set_autostart;
 use crate::utils::color::*;
 use crate::utils::font::{DrawTextCachedParams, FontManager};
 use crate::utils::icon::get_app_icon;
@@ -23,6 +21,8 @@ use winit::event_loop::{ActiveEventLoop, EventLoop};
 use winit::keyboard::{Key, NamedKey};
 use winit::raw_window_handle::{HasWindowHandle, RawWindowHandle};
 use winit::window::{Window, WindowButtons, WindowId};
+
+pub mod input;
 
 const WIN_W: f32 = 666.0;
 const WIN_H: f32 = 666.0;
@@ -266,7 +266,7 @@ impl SettingsApp {
             unsafe {
                 let _ = DwmSetWindowAttribute(
                     hwnd,
-                    DWMWINDOWATTRIBUTE(20), // DWMWA_USE_IMMERSIVE_DARK_MODE
+                    DWMWINDOWATTRIBUTE(20),
                     &use_dark as *const _ as *const _,
                     std::mem::size_of::<i32>() as u32,
                 );
@@ -699,8 +699,6 @@ impl SettingsApp {
     fn get_monitor_list(&self) -> Vec<String> {
         use windows::Win32::Graphics::Gdi::*;
         let mut monitors: Vec<String> = Vec::new();
-        // SAFETY: EnumDisplayDevicesW reads display device info. We provide a zeroed
-        // DISPLAY_DEVICEW with correct cb size. idx increments safely. No mutable global state.
         unsafe {
             let mut idx = 0u32;
             let mut active_count = 0;
@@ -1216,682 +1214,6 @@ impl SettingsApp {
             _ => SwitchAnimator::new(&[]),
         }
     }
-
-    fn handle_click(&mut self) {
-        let (mx, my) = self.logical_mouse_pos;
-
-        if let Some(popup) = &self.popup {
-            if let Some(i) = popup.hit_test_item(mx, my) {
-                let value = popup.values[i].clone();
-                match popup.kind {
-                    PopupKind::LyricsSource => {
-                        self.config.lyrics_source = value;
-                    }
-                    PopupKind::Language => {
-                        self.config.language = value;
-                        set_lang(&self.config.language);
-                    }
-                    PopupKind::Monitor => {
-                        self.config.monitor_index = value.parse::<i32>().unwrap_or(0);
-                    }
-                    PopupKind::IslandStyle => {
-                        self.config.island_style = value;
-                    }
-                    PopupKind::DockPositionPopup => {
-                        self.config.dock_position = value
-                            .parse::<DockPosition>()
-                            .unwrap_or(DockPosition::TopCenter);
-                    }
-                    PopupKind::SettingsTheme => {
-                        self.config.settings_theme = value.clone();
-                        self.update_theme();
-                    }
-                    PopupKind::MiniCoverShape => {
-                        self.config.mini_cover_shape = value;
-                    }
-                    PopupKind::ExpandedCoverShape => {
-                        self.config.expanded_cover_shape = value;
-                    }
-                }
-                save_config(&self.config);
-                self.mark_items_dirty();
-            }
-            self.popup = None;
-            self.anim.set_with_speed(POPUP_OPACITY_KEY, 0.0, 0.3);
-            if let Some(win) = &self.window {
-                win.request_redraw();
-            }
-            return;
-        }
-
-        if mx < SIDEBAR_W {
-            let pages = 3;
-            let start_y = 20.0;
-            for i in 0..pages {
-                let row_y = start_y + i as f32 * (SIDEBAR_ROW_H + 2.0);
-                if my >= row_y
-                    && my <= row_y + SIDEBAR_ROW_H
-                    && (SIDEBAR_PAD..=SIDEBAR_W - SIDEBAR_PAD).contains(&mx)
-                {
-                    if self.active_page != i as usize {
-                        self.active_page = i as usize;
-                        self.scroll_y = 0.0;
-                        self.target_scroll_y = 0.0;
-                        self.scroll_vel_y = 0.0;
-                        self.mark_items_dirty();
-                        if let Some(win) = &self.window {
-                            win.request_redraw();
-                        }
-                    }
-                    return;
-                }
-            }
-            return;
-        }
-
-        let scale = self
-            .window
-            .as_ref()
-            .map(|w| w.scale_factor() as f32)
-            .unwrap_or(1.0);
-        let content_w = self.win_w / scale - SIDEBAR_W;
-
-        if self.active_page == 0 && (SUB_TAB_START_Y..=SUB_TAB_START_Y + SUB_TAB_H).contains(&my) {
-            let tabs = [
-                tr("section_appearance"),
-                tr("section_effects"),
-                tr("section_behavior"),
-            ];
-            let tab_count = tabs.len();
-            let tab_w = content_w / tab_count as f32;
-            let rel_x = mx - SIDEBAR_W;
-            let tab_idx = (rel_x / tab_w) as usize;
-            if tab_idx < tab_count && self.active_sub_page != tab_idx {
-                self.active_sub_page = tab_idx;
-                self.scroll_y = 0.0;
-                self.target_scroll_y = 0.0;
-                self.scroll_vel_y = 0.0;
-                self.mark_items_dirty();
-                if let Some(win) = &self.window {
-                    win.request_redraw();
-                }
-            }
-            return;
-        }
-
-        let content_x = mx - SIDEBAR_W;
-        let content_start_y = if self.active_page == 0 {
-            SUB_TAB_START_Y + SUB_TAB_H + CONTENT_START_Y
-        } else {
-            CONTENT_START_Y
-        };
-        let content_y = my + self.scroll_y;
-        let items = self.build_current_items();
-
-        match self.active_page {
-            0 => {
-                self.handle_general_click(&items, content_x, content_y, content_w, content_start_y)
-            }
-            1 => self.handle_music_click(&items, content_x, content_y, content_w, content_start_y),
-            2 => self.handle_about_click(&items, content_x, content_y, content_w, content_start_y),
-            _ => {}
-        }
-    }
-
-    fn handle_general_click(
-        &mut self,
-        items: &[SettingsItem],
-        mx: f32,
-        my: f32,
-        width: f32,
-        start_y: f32,
-    ) {
-        let scale = self
-            .window
-            .as_ref()
-            .map(|w| w.scale_factor() as f32)
-            .unwrap_or(1.0);
-        let result = hit_test(items, mx, my, start_y, width);
-        let mut changed = false;
-
-        match result {
-            ClickResult::StepperDec(idx) | ClickResult::StepperInc(idx) => {
-                let is_dec = matches!(result, ClickResult::StepperDec(_));
-                if let Some(item) = items.get(idx)
-                    && let SettingsItem::RowStepper { label, .. } = item
-                {
-                    let l = label.clone();
-                    if l == tr("global_scale") {
-                        if is_dec {
-                            self.config.global_scale =
-                                ((self.config.global_scale - 0.05) * 100.0).round() / 100.0;
-                            self.config.global_scale = self.config.global_scale.max(0.5);
-                        } else {
-                            self.config.global_scale =
-                                ((self.config.global_scale + 0.05) * 100.0).round() / 100.0;
-                            self.config.global_scale = self.config.global_scale.min(5.0);
-                        }
-                        changed = true;
-                    } else if l == tr("base_width") {
-                        if is_dec {
-                            self.config.base_width = (self.config.base_width - 5.0).max(40.0);
-                        } else {
-                            self.config.base_width += 5.0;
-                        }
-                        changed = true;
-                    } else if l == tr("base_height") {
-                        if is_dec {
-                            self.config.base_height = (self.config.base_height - 2.0).max(15.0);
-                        } else {
-                            self.config.base_height += 2.0;
-                        }
-                        changed = true;
-                    } else if l == tr("expanded_width") {
-                        if is_dec {
-                            self.config.expanded_width =
-                                (self.config.expanded_width - 10.0).max(200.0);
-                        } else {
-                            self.config.expanded_width += 10.0;
-                        }
-                        changed = true;
-                    } else if l == tr("expanded_height") {
-                        if is_dec {
-                            self.config.expanded_height =
-                                (self.config.expanded_height - 10.0).max(100.0);
-                        } else {
-                            self.config.expanded_height += 10.0;
-                        }
-                        changed = true;
-                    } else if l == tr("position_x_offset") {
-                        if is_dec {
-                            self.config.position_x_offset -= 5;
-                        } else {
-                            self.config.position_x_offset += 5;
-                        }
-                        changed = true;
-                    } else if l == tr("position_y_offset") {
-                        if is_dec {
-                            self.config.position_y_offset -= 5;
-                        } else {
-                            self.config.position_y_offset += 5;
-                        }
-                        changed = true;
-                    } else if l == tr("font_size") {
-                        if is_dec {
-                            self.config.font_size = (self.config.font_size - 1.0).max(0.0);
-                        } else {
-                            self.config.font_size = (self.config.font_size + 1.0).min(30.0);
-                        }
-                        changed = true;
-                    } else if l == tr("hide_delay") {
-                        if is_dec {
-                            self.config.auto_hide_delay =
-                                (self.config.auto_hide_delay - 1.0).max(1.0);
-                        } else {
-                            self.config.auto_hide_delay =
-                                (self.config.auto_hide_delay + 1.0).min(60.0);
-                        }
-                        changed = true;
-                    } else if l == tr("update_interval") {
-                        if is_dec {
-                            self.config.update_check_interval =
-                                (self.config.update_check_interval - 1.0).max(1.0);
-                        } else {
-                            self.config.update_check_interval =
-                                (self.config.update_check_interval + 1.0).min(24.0);
-                        }
-                        changed = true;
-                    }
-                }
-            }
-            ClickResult::Switch(idx) => {
-                let label = items
-                    .iter()
-                    .filter_map(|item| match item {
-                        SettingsItem::RowSwitch { label, .. } => Some(label.clone()),
-                        _ => None,
-                    })
-                    .nth(idx);
-                if let Some(label) = label {
-                    match label.as_str() {
-                        l if l == tr("adaptive_border") => {
-                            self.config.adaptive_border = !self.config.adaptive_border
-                        }
-                        l if l == tr("motion_blur") => {
-                            self.config.motion_blur = !self.config.motion_blur
-                        }
-                        l if l == tr("cover_rotate") => {
-                            self.config.cover_rotate = !self.config.cover_rotate
-                        }
-                        l if l == tr("audio_gate") => {
-                            self.config.audio_gate = !self.config.audio_gate;
-                        }
-                        l if l == tr("auto_gate") => self.config.auto_gate = !self.config.auto_gate,
-                        l if l == tr("mini_controls") => {
-                            self.config.mini_controls = !self.config.mini_controls
-                        }
-                        l if l == tr("start_boot") => {
-                            self.config.auto_start = !self.config.auto_start;
-                            let _ = set_autostart(self.config.auto_start);
-                        }
-                        l if l == tr("auto_hide") => self.config.auto_hide = !self.config.auto_hide,
-                        l if l == tr("check_updates") => {
-                            self.config.check_for_updates = !self.config.check_for_updates
-                        }
-                        _ => {
-                            log::warn!("WinIsland: unhandled switch label: {}", label);
-                        }
-                    }
-                }
-                self.sync_switch_targets();
-                changed = true;
-            }
-            ClickResult::FontSelect(_) => {
-                if let Some(path) = rfd::FileDialog::new()
-                    .add_filter("Fonts", &["ttf", "otf"])
-                    .pick_file()
-                {
-                    self.config.custom_font_path = Some(path.to_string_lossy().into_owned());
-                    FontManager::global().refresh_custom_font();
-                    changed = true;
-                }
-            }
-            ClickResult::FontReset(_) => {
-                self.config.custom_font_path = None;
-                FontManager::global().refresh_custom_font();
-                changed = true;
-            }
-            ClickResult::SourceButton(idx) => {
-                let content_w = width;
-                let mut btn_content_y = start_y;
-                for item in items.iter().take(idx) {
-                    btn_content_y += item.height();
-                }
-                let cy = btn_content_y + ROW_HEIGHT / 2.0;
-                let btn_x = SIDEBAR_W + CONTENT_PADDING + content_w - GROUP_INNER_PAD - POPUP_BTN_W;
-                let btn_y = cy - POPUP_BTN_H / 2.0 - self.scroll_y;
-
-                if let Some(SettingsItem::RowSourceSelect { label, .. }) = items.get(idx) {
-                    if label == &tr("monitor") {
-                        let monitors = self.get_monitor_list();
-                        let selected_idx = (self.config.monitor_index as usize)
-                            .min(monitors.len().saturating_sub(1));
-                        let values: Vec<String> =
-                            (0..monitors.len()).map(|i| i.to_string()).collect();
-                        self.popup = Some(PopupState::new(
-                            PopupKind::Monitor,
-                            Rect::from_xywh(btn_x, btn_y, POPUP_BTN_W, POPUP_BTN_H),
-                            monitors,
-                            values,
-                            selected_idx,
-                            self.win_w / scale,
-                            self.win_h / scale,
-                        ));
-                    } else if label == &tr("island_style") {
-                        let selected_idx = match self.config.island_style.as_str() {
-                            "glass" => 1,
-                            "mica" => 2,
-                            "dynamic" => 3,
-                            "liquid_glass" => 4,
-                            _ => 0,
-                        };
-                        self.popup = Some(PopupState::new(
-                            PopupKind::IslandStyle,
-                            Rect::from_xywh(btn_x, btn_y, POPUP_BTN_W, POPUP_BTN_H),
-                            vec![
-                                tr("style_default"),
-                                tr("style_glass"),
-                                tr("style_mica"),
-                                tr("style_dynamic"),
-                                tr("style_liquid_glass"),
-                            ],
-                            vec![
-                                "default".to_string(),
-                                "glass".to_string(),
-                                "mica".to_string(),
-                                "dynamic".to_string(),
-                                "liquid_glass".to_string(),
-                            ],
-                            selected_idx,
-                            self.win_w / scale,
-                            self.win_h / scale,
-                        ));
-                    } else if label == &tr("dock_position") {
-                        let dp = self.config.dock_position;
-                        let selected_idx = match dp {
-                            DockPosition::TopCenter => 0,
-                            DockPosition::TopLeft => 1,
-                            DockPosition::TopRight => 2,
-                            DockPosition::BottomCenter => 3,
-                            DockPosition::BottomLeft => 4,
-                            DockPosition::BottomRight => 5,
-                        };
-                        self.popup = Some(PopupState::new(
-                            PopupKind::DockPositionPopup,
-                            Rect::from_xywh(btn_x, btn_y, POPUP_BTN_W, POPUP_BTN_H),
-                            vec![
-                                tr("dock_position_top_center"),
-                                tr("dock_position_top_left"),
-                                tr("dock_position_top_right"),
-                                tr("dock_position_bottom_center"),
-                                tr("dock_position_bottom_left"),
-                                tr("dock_position_bottom_right"),
-                            ],
-                            vec![
-                                "top_center".to_string(),
-                                "top_left".to_string(),
-                                "top_right".to_string(),
-                                "bottom_center".to_string(),
-                                "bottom_left".to_string(),
-                                "bottom_right".to_string(),
-                            ],
-                            selected_idx,
-                            self.win_w / scale,
-                            self.win_h / scale,
-                        ));
-                    } else if label == &tr("settings_theme") {
-                        let selected_idx = match self.config.settings_theme.as_str() {
-                            "light" => 1,
-                            "dark" => 2,
-                            _ => 0,
-                        };
-                        self.popup = Some(PopupState::new(
-                            PopupKind::SettingsTheme,
-                            Rect::from_xywh(btn_x, btn_y, POPUP_BTN_W, POPUP_BTN_H),
-                            vec![tr("theme_system"), tr("theme_light"), tr("theme_dark")],
-                            vec![
-                                "system".to_string(),
-                                "light".to_string(),
-                                "dark".to_string(),
-                            ],
-                            selected_idx,
-                            self.win_w / scale,
-                            self.win_h / scale,
-                        ));
-                    } else if label == &tr("mini_cover_shape") {
-                        let selected_idx = if self.config.mini_cover_shape == "circle" {
-                            1
-                        } else {
-                            0
-                        };
-                        self.popup = Some(PopupState::new(
-                            PopupKind::MiniCoverShape,
-                            Rect::from_xywh(btn_x, btn_y, POPUP_BTN_W, POPUP_BTN_H),
-                            vec![tr("shape_square"), tr("shape_circle")],
-                            vec!["square".to_string(), "circle".to_string()],
-                            selected_idx,
-                            self.win_w / scale,
-                            self.win_h / scale,
-                        ));
-                    } else if label == &tr("expanded_cover_shape") {
-                        let selected_idx = if self.config.expanded_cover_shape == "circle" {
-                            1
-                        } else {
-                            0
-                        };
-                        self.popup = Some(PopupState::new(
-                            PopupKind::ExpandedCoverShape,
-                            Rect::from_xywh(btn_x, btn_y, POPUP_BTN_W, POPUP_BTN_H),
-                            vec![tr("shape_square"), tr("shape_circle")],
-                            vec!["square".to_string(), "circle".to_string()],
-                            selected_idx,
-                            self.win_w / scale,
-                            self.win_h / scale,
-                        ));
-                    } else {
-                        let lang = current_lang();
-                        self.popup = Some(PopupState::new(
-                            PopupKind::Language,
-                            Rect::from_xywh(btn_x, btn_y, POPUP_BTN_W, POPUP_BTN_H),
-                            vec!["English".to_string(), "中文".to_string()],
-                            vec!["en".to_string(), "zh".to_string()],
-                            if lang == "zh" { 1 } else { 0 },
-                            self.win_w / scale,
-                            self.win_h / scale,
-                        ));
-                    }
-                    self.anim.set_with_speed(POPUP_OPACITY_KEY, 1.0, 0.25);
-                    if let Some(win) = &self.window {
-                        win.request_redraw();
-                    }
-                }
-            }
-            ClickResult::CenterLink(_) => {
-                self.config = AppConfig::default();
-                init_i18n(&self.config.language);
-                FontManager::global().refresh_custom_font();
-                self.switch_anim = SwitchAnimator::new(&[
-                    self.config.adaptive_border,
-                    self.config.motion_blur,
-                    self.config.cover_rotate,
-                    self.config.audio_gate,
-                    self.config.auto_gate,
-                    self.config.mini_controls,
-                    self.config.auto_start,
-                    self.config.auto_hide,
-                    self.config.check_for_updates,
-                    self.config.smtc_enabled,
-                    self.config.show_lyrics,
-                    self.config.lyrics_fallback,
-                    self.config.lyrics_scroll,
-                ]);
-                changed = true;
-            }
-            _ => {}
-        }
-
-        if changed {
-            self.mark_items_dirty();
-            save_config(&self.config);
-            if let Some(win) = &self.window {
-                win.request_redraw();
-            }
-        }
-    }
-
-    fn handle_music_click(
-        &mut self,
-        items: &[SettingsItem],
-        mx: f32,
-        my: f32,
-        width: f32,
-        start_y: f32,
-    ) {
-        let scale = self
-            .window
-            .as_ref()
-            .map(|w| w.scale_factor() as f32)
-            .unwrap_or(1.0);
-        let result = hit_test(items, mx, my, start_y, width);
-        let mut changed = false;
-
-        match result {
-            ClickResult::Switch(idx) => {
-                let label = items
-                    .iter()
-                    .filter_map(|item| match item {
-                        SettingsItem::RowSwitch { label, .. } => Some(label.clone()),
-                        _ => None,
-                    })
-                    .nth(idx);
-                if let Some(label) = label {
-                    match label.as_str() {
-                        l if l == tr("smtc_control") => {
-                            self.config.smtc_enabled = !self.config.smtc_enabled
-                        }
-                        l if l == tr("show_lyrics") => {
-                            self.config.show_lyrics = !self.config.show_lyrics
-                        }
-                        l if l == tr("lyrics_fallback") => {
-                            if self.config.show_lyrics {
-                                self.config.lyrics_fallback = !self.config.lyrics_fallback
-                            }
-                        }
-                        l if l == tr("lyrics_scroll") => {
-                            if self.config.show_lyrics {
-                                self.config.lyrics_scroll = !self.config.lyrics_scroll
-                            }
-                        }
-                        _ => {
-                            log::warn!("WinIsland: unhandled switch label: {}", label);
-                        }
-                    }
-                }
-                self.sync_switch_targets();
-                changed = true;
-            }
-            ClickResult::SourceButton(idx) => {
-                let content_w = width;
-                let mut btn_content_y = start_y;
-                for item in items.iter().take(idx) {
-                    btn_content_y += item.height();
-                }
-                let cy = btn_content_y + ROW_HEIGHT / 2.0;
-                let btn_x = SIDEBAR_W + CONTENT_PADDING + content_w - GROUP_INNER_PAD - POPUP_BTN_W;
-                let btn_y = cy - POPUP_BTN_H / 2.0 - self.scroll_y;
-
-                let source = &self.config.lyrics_source;
-                self.popup = Some(PopupState::new(
-                    PopupKind::LyricsSource,
-                    Rect::from_xywh(btn_x, btn_y, POPUP_BTN_W, POPUP_BTN_H),
-                    vec!["163".to_string(), "LRCLIB".to_string()],
-                    vec!["163".to_string(), "lrclib".to_string()],
-                    if source == "163" { 0 } else { 1 },
-                    self.win_w / scale,
-                    self.win_h / scale,
-                ));
-                self.anim.set_with_speed(POPUP_OPACITY_KEY, 1.0, 0.25);
-                if let Some(win) = &self.window {
-                    win.request_redraw();
-                }
-            }
-            ClickResult::StepperDec(idx) | ClickResult::StepperInc(idx) => {
-                let is_dec = matches!(result, ClickResult::StepperDec(_));
-                if let Some(item) = items.get(idx)
-                    && let SettingsItem::RowStepper { label, .. } = item
-                {
-                    if label == &tr("lyrics_delay") && self.config.show_lyrics {
-                        if is_dec {
-                            self.config.lyrics_delay =
-                                ((self.config.lyrics_delay * 10.0 - 1.0).round() / 10.0).max(-10.0);
-                        } else {
-                            self.config.lyrics_delay =
-                                ((self.config.lyrics_delay * 10.0 + 1.0).round() / 10.0).min(10.0);
-                        }
-                        changed = true;
-                    } else if label == &tr("lyrics_scroll_max_width")
-                        && self.config.show_lyrics
-                        && self.config.lyrics_scroll
-                    {
-                        if is_dec {
-                            self.config.lyrics_scroll_max_width =
-                                (self.config.lyrics_scroll_max_width - 10.0).max(100.0);
-                        } else {
-                            self.config.lyrics_scroll_max_width =
-                                (self.config.lyrics_scroll_max_width + 10.0).min(500.0);
-                        }
-                        changed = true;
-                    }
-                }
-            }
-            ClickResult::AppItem(idx)
-                if self.config.smtc_enabled && !self.detected_apps.is_empty() =>
-            {
-                let app_start = items
-                    .iter()
-                    .position(|i| matches!(i, SettingsItem::RowAppItem { .. }))
-                    .unwrap_or(items.len());
-                let app_idx = idx - app_start;
-                if app_idx < self.detected_apps.len() {
-                    let app = &self.detected_apps[app_idx];
-                    if self.config.smtc_apps.contains(app) {
-                        self.config.smtc_apps.retain(|a| a != app);
-                    } else {
-                        self.config.smtc_apps.push(app.clone());
-                        if !self.config.smtc_known_apps.contains(app) {
-                            self.config.smtc_known_apps.push(app.clone());
-                        }
-                    }
-                    changed = true;
-                }
-            }
-            _ => {}
-        }
-
-        if changed {
-            self.mark_items_dirty();
-            save_config(&self.config);
-            if let Some(win) = &self.window {
-                win.request_redraw();
-            }
-        }
-    }
-
-    fn handle_about_click(
-        &mut self,
-        items: &[SettingsItem],
-        mx: f32,
-        my: f32,
-        width: f32,
-        start_y: f32,
-    ) {
-        let result = hit_test(items, mx, my, start_y, width);
-        if let ClickResult::CenterLink(_) = result {
-            let _ = open::that(APP_HOMEPAGE);
-        }
-    }
-
-    fn get_hover_state(&mut self) -> bool {
-        let (mx, my) = self.logical_mouse_pos;
-
-        if let Some(popup) = &self.popup {
-            let menu = popup.menu_rect();
-            if mx >= menu.left && mx <= menu.right && my >= menu.top && my <= menu.bottom {
-                return true;
-            }
-        }
-
-        if mx < SIDEBAR_W {
-            let start_y = 20.0;
-            for i in 0..3 {
-                let row_y = start_y + i as f32 * (SIDEBAR_ROW_H + 2.0);
-                if my >= row_y
-                    && my <= row_y + SIDEBAR_ROW_H
-                    && (SIDEBAR_PAD..=SIDEBAR_W - SIDEBAR_PAD).contains(&mx)
-                {
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        let scale = self
-            .window
-            .as_ref()
-            .map(|w| w.scale_factor() as f32)
-            .unwrap_or(1.0);
-        let content_w = self.win_w / scale - SIDEBAR_W;
-
-        if self.active_page == 0 && (SUB_TAB_START_Y..=SUB_TAB_START_Y + SUB_TAB_H).contains(&my) {
-            return true;
-        }
-
-        let content_x = mx - SIDEBAR_W;
-        let content_start_y = if self.active_page == 0 {
-            SUB_TAB_START_Y + SUB_TAB_H + CONTENT_START_Y
-        } else {
-            CONTENT_START_Y
-        };
-        let content_y = my + self.scroll_y;
-        self.ensure_items_cache();
-        hover_test(
-            &self.cached_items,
-            content_x,
-            content_y,
-            content_start_y,
-            content_w,
-        )
-    }
 }
 
 impl ApplicationHandler for SettingsApp {
@@ -2195,8 +1517,6 @@ impl ApplicationHandler for SettingsApp {
         let frame_start = Instant::now();
         self.frame_count += 1;
         if self.frame_count.is_multiple_of(30) {
-            // SAFETY: OpenMutexW opens an existing named mutex. The mutex name is a static
-            // string literal. CloseHandle is called on the valid handle returned by OpenMutexW.
             unsafe {
                 let h = OpenMutexW(
                     MUTEX_ALL_ACCESS,
